refactor: moves proxies APIs to a separate module
